+++ packages/builder/builder-shared/src/types/config/security.ts
@@ -5,5 +5,9 @@ export type SriOptions = {
 };
 
 export interface SharedSecurityConfig {
+  /**
+   * Adding an integrity attribute (`integrity`) to sub-resources introduced by HTML allows the browser to
+   * verify the integrity of the introduced resource, thus preventing tampering with the downloaded resource.
+   */
   sri?: SriOptions | boolean;
 }