+++ packages/cli/plugin-changeset/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-changeset
 
+## 1.21.0
+
+### Patch Changes
+
+- @modern-js/utils@1.21.0
+- @modern-js/plugin-i18n@1.21.0
+
 ## 1.20.1
 
 ### Patch Changes